feat: ConnectionMenu own file
